@@ -19,44 +19,43 @@
 
 package top.theillusivec4.culinaryconstruct.client;
 
-import com.mojang.blaze3d.matrix.MatrixStack;
 import com.mojang.blaze3d.systems.RenderSystem;
+import com.mojang.blaze3d.vertex.PoseStack;
 import javax.annotation.Nonnull;
 import net.minecraft.client.Minecraft;
-import net.minecraft.client.gui.screen.inventory.ContainerScreen;
-import net.minecraft.client.gui.widget.TextFieldWidget;
-import net.minecraft.entity.player.PlayerInventory;
-import net.minecraft.inventory.container.Container;
-import net.minecraft.inventory.container.IContainerListener;
-import net.minecraft.item.ItemStack;
-import net.minecraft.util.NonNullList;
-import net.minecraft.util.ResourceLocation;
-import net.minecraft.util.text.ITextComponent;
-import net.minecraft.util.text.TranslationTextComponent;
+import net.minecraft.client.gui.components.EditBox;
+import net.minecraft.client.gui.screens.inventory.AbstractContainerScreen;
+import net.minecraft.network.chat.Component;
+import net.minecraft.network.chat.TranslatableComponent;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.entity.player.Inventory;
+import net.minecraft.world.inventory.AbstractContainerMenu;
+import net.minecraft.world.inventory.ContainerListener;
+import net.minecraft.world.item.ItemStack;
 import top.theillusivec4.culinaryconstruct.CulinaryConstruct;
 import top.theillusivec4.culinaryconstruct.common.inventory.CulinaryStationContainer;
 import top.theillusivec4.culinaryconstruct.common.network.CPacketRename;
 import top.theillusivec4.culinaryconstruct.common.network.CulinaryConstructNetwork;
 
-public class CulinaryScreen extends ContainerScreen<CulinaryStationContainer> implements
-    IContainerListener {
+public class CulinaryScreen extends AbstractContainerScreen<CulinaryStationContainer> implements
+    ContainerListener {
 
   private static final ResourceLocation GUI_BACKGROUND = new ResourceLocation(
-      CulinaryConstruct.MODID, "textures/gui/culinary_station_gui.png");
+      CulinaryConstruct.MOD_ID, "textures/gui/culinary_station_gui.png");
   private static final int WIDTH = 176;
   private static final int HEIGHT = 161;
 
-  private TextFieldWidget nameField;
+  private EditBox nameField;
 
-  public CulinaryScreen(CulinaryStationContainer screenContainer, PlayerInventory inv,
-      ITextComponent titleIn) {
+  public CulinaryScreen(CulinaryStationContainer screenContainer, Inventory inv,
+                        Component titleIn) {
     super(screenContainer, inv, titleIn);
-    this.xSize = WIDTH;
-    this.ySize = HEIGHT;
-    this.titleX = 60;
-    this.titleY = 6;
-    this.playerInventoryTitleX = 8;
-    this.playerInventoryTitleY = 67;
+    this.imageWidth = WIDTH;
+    this.imageHeight = HEIGHT;
+    this.titleLabelX = 60;
+    this.titleLabelY = 6;
+    this.inventoryLabelX = 8;
+    this.inventoryLabelY = 67;
   }
 
   @Override
@@ -64,37 +63,37 @@ public class CulinaryScreen extends ContainerScreen<CulinaryStationContainer> im
     super.init();
 
     if (this.minecraft != null) {
-      this.minecraft.keyboardListener.enableRepeatEvents(true);
+      this.minecraft.keyboardHandler.setSendRepeatsToGui(true);
     }
-    int i = (this.width - this.xSize) / 2;
-    int j = (this.height - this.ySize) / 2;
-    this.nameField = new TextFieldWidget(this.font, i + 62, j + 20, 103, 12,
-        new TranslationTextComponent("culinaryconstruct.culinary_container"));
+    int i = (this.width - this.imageWidth) / 2;
+    int j = (this.height - this.imageHeight) / 2;
+    this.nameField = new EditBox(this.font, i + 62, j + 20, 103, 12,
+        new TranslatableComponent("culinaryconstruct.culinary_container"));
     this.nameField.setCanLoseFocus(false);
     this.nameField.setTextColor(-1);
-    this.nameField.setDisabledTextColour(-1);
-    this.nameField.setEnableBackgroundDrawing(false);
-    this.nameField.setMaxStringLength(35);
+    this.nameField.setTextColorUneditable(-1);
+    this.nameField.setBordered(false);
+    this.nameField.setMaxLength(35);
     this.nameField.setResponder(this::updateName);
-    this.children.add(this.nameField);
-    this.setFocusedDefault(this.nameField);
-    this.container.addListener(this);
+    this.addWidget(this.nameField);
+    this.setInitialFocus(this.nameField);
+    this.menu.addSlotListener(this);
   }
 
   @Override
   public void resize(@Nonnull Minecraft minecraft, int mouse1, int mouse2) {
-    String s = this.nameField.getText();
+    String s = this.nameField.getValue();
     this.init(minecraft, mouse1, mouse2);
-    this.nameField.setText(s);
+    this.nameField.setValue(s);
   }
 
   @Override
-  public void onClose() {
-    super.onClose();
-    this.container.removeListener(this);
+  public void removed() {
+    super.removed();
+    this.menu.removeSlotListener(this);
 
     if (this.minecraft != null) {
-      this.minecraft.keyboardListener.enableRepeatEvents(false);
+      this.minecraft.keyboardHandler.setSendRepeatsToGui(false);
     }
   }
 
@@ -102,24 +101,24 @@ public class CulinaryScreen extends ContainerScreen<CulinaryStationContainer> im
   public boolean keyPressed(int key1, int key2, int key3) {
 
     if (key1 == 256 && this.minecraft != null && this.minecraft.player != null) {
-      this.minecraft.player.closeScreen();
+      this.minecraft.player.closeContainer();
     }
-    return this.nameField.keyPressed(key1, key2, key3) || this.nameField.canWrite() || super
+    return this.nameField.keyPressed(key1, key2, key3) || this.nameField.canConsumeInput() || super
         .keyPressed(key1, key2, key3);
   }
 
   @Override
-  public void render(@Nonnull MatrixStack matrixStack, int mouseX, int mouseY, float partialTicks) {
+  public void render(@Nonnull PoseStack matrixStack, int mouseX, int mouseY, float partialTicks) {
     this.renderBackground(matrixStack);
     super.render(matrixStack, mouseX, mouseY, partialTicks);
     RenderSystem.disableBlend();
     this.nameField.render(matrixStack, mouseX, mouseY, partialTicks);
-    this.renderHoveredTooltip(matrixStack, mouseX, mouseY);
+    this.renderTooltip(matrixStack, mouseX, mouseY);
   }
 
   private void updateName(String name) {
-    if (this.container.getSlot(6).getHasStack()) {
-      this.container.updateItemName(name);
+    if (this.menu.getSlot(6).hasItem()) {
+      this.menu.updateItemName(name);
 
       if (this.minecraft != null) {
         CulinaryConstructNetwork.INSTANCE.sendToServer(new CPacketRename(name));
@@ -127,45 +126,39 @@ public class CulinaryScreen extends ContainerScreen<CulinaryStationContainer> im
     }
   }
 
-  @SuppressWarnings("deprecation")
   @Override
-  protected void drawGuiContainerBackgroundLayer(@Nonnull MatrixStack matrixStack,
-      float partialTicks, int mouseX, int mouseY) {
+  protected void renderBg(@Nonnull PoseStack matrixStack,
+                          float partialTicks, int mouseX, int mouseY) {
 
     if (this.minecraft != null) {
-      RenderSystem.color4f(1.0F, 1.0F, 1.0F, 1.0F);
-      this.minecraft.getTextureManager().bindTexture(GUI_BACKGROUND);
-      int i = (this.width - this.xSize) / 2;
-      int j = (this.height - this.ySize) / 2;
-      this.blit(matrixStack, i, j, 0, 0, this.xSize, this.ySize);
+      RenderSystem.setShaderColor(1.0F, 1.0F, 1.0F, 1.0F);
+      RenderSystem.setShaderTexture(0, GUI_BACKGROUND);
+      int i = (this.width - this.imageWidth) / 2;
+      int j = (this.height - this.imageHeight) / 2;
+      this.blit(matrixStack, i, j, 0, 0, this.imageWidth, this.imageHeight);
       this.blit(matrixStack, i + 59, j + 16, 0,
-          this.ySize + (this.container.getSlot(0).getHasStack() ? 0 : 16), 110, 16);
+          this.imageHeight + (this.menu.getSlot(0).hasItem() ? 0 : 16), 110, 16);
 
-      if (this.container.getSlot(0).getHasStack() && !this.container.getSlot(6).getHasStack()) {
-        this.blit(matrixStack, i + 133, j + 43, this.xSize, 0, 18, 18);
+      if (this.menu.getSlot(0).hasItem() && !this.menu.getSlot(6).hasItem()) {
+        this.blit(matrixStack, i + 133, j + 43, this.imageWidth, 0, 18, 18);
       }
     }
   }
 
   @Override
-  public void sendAllContents(@Nonnull Container containerToSend,
-      @Nonnull NonNullList<ItemStack> itemsList) {
-    this.sendSlotContents(containerToSend, 6, containerToSend.getSlot(6).getStack());
-  }
-
-  @Override
-  public void sendSlotContents(@Nonnull Container containerToSend, int slotInd,
-      @Nonnull ItemStack stack) {
+  public void slotChanged(@Nonnull AbstractContainerMenu containerToSend, int slotInd,
+                          @Nonnull ItemStack stack) {
 
     if (slotInd == 6) {
-      this.nameField.setText(stack.isEmpty() ? "" : this.nameField.getText());
-      this.nameField.setEnabled(!stack.isEmpty());
-      this.setListener(this.nameField);
+      this.nameField.setValue(stack.isEmpty() ? "" : this.nameField.getValue());
+      this.nameField.setEditable(!stack.isEmpty());
+      this.setFocused(this.nameField);
     }
   }
 
   @Override
-  public void sendWindowProperty(@Nonnull Container containerIn, int varToUpdate, int newValue) {
-    //NO-OP
+  public void dataChanged(@Nonnull AbstractContainerMenu pContainerMenu, int pDataSlotIndex,
+                          int pValue) {
+    this.slotChanged(pContainerMenu, 6, pContainerMenu.getSlot(6).getItem());
   }
 }
