@@ -3,7 +3,7 @@ package top.theillusivec4.culinaryconstruct.client.model;
 import com.google.gson.JsonDeserializationContext;
 import com.google.gson.JsonObject;
 import javax.annotation.Nonnull;
-import net.minecraft.resources.IResourceManager;
+import net.minecraft.server.packs.resources.ResourceManager;
 import net.minecraftforge.client.model.IModelLoader;
 
 public enum FoodBowlLoader implements IModelLoader<FoodBowlModel> {
@@ -11,7 +11,7 @@ public enum FoodBowlLoader implements IModelLoader<FoodBowlModel> {
   INSTANCE;
 
   @Override
-  public void onResourceManagerReload(@Nonnull IResourceManager resourceManager) {
+  public void onResourceManagerReload(@Nonnull ResourceManager resourceManager) {
 
   }
 
