@@ -22,11 +22,11 @@ package top.theillusivec4.culinaryconstruct.common.network;
 import java.util.function.BiConsumer;
 import java.util.function.Function;
 import java.util.function.Supplier;
-import net.minecraft.network.PacketBuffer;
-import net.minecraft.util.ResourceLocation;
-import net.minecraftforge.fml.network.NetworkEvent.Context;
-import net.minecraftforge.fml.network.NetworkRegistry;
-import net.minecraftforge.fml.network.simple.SimpleChannel;
+import net.minecraft.network.FriendlyByteBuf;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraftforge.fmllegacy.network.NetworkEvent;
+import net.minecraftforge.fmllegacy.network.NetworkRegistry;
+import net.minecraftforge.fmllegacy.network.simple.SimpleChannel;
 import top.theillusivec4.culinaryconstruct.CulinaryConstruct;
 
 public class CulinaryConstructNetwork {
@@ -38,7 +38,7 @@ public class CulinaryConstructNetwork {
 
   public static void register() {
     INSTANCE = NetworkRegistry.ChannelBuilder
-        .named(new ResourceLocation(CulinaryConstruct.MODID, "main"))
+        .named(new ResourceLocation(CulinaryConstruct.MOD_ID, "main"))
         .networkProtocolVersion(() -> PTC_VERSION).clientAcceptedVersions(PTC_VERSION::equals)
         .serverAcceptedVersions(PTC_VERSION::equals).simpleChannel();
 
@@ -46,8 +46,9 @@ public class CulinaryConstructNetwork {
         CPacketRename::handle);
   }
 
-  private static <M> void register(Class<M> messageType, BiConsumer<M, PacketBuffer> encoder,
-      Function<PacketBuffer, M> decoder, BiConsumer<M, Supplier<Context>> messageConsumer) {
+  private static <M> void register(Class<M> messageType, BiConsumer<M, FriendlyByteBuf> encoder,
+                                   Function<FriendlyByteBuf, M> decoder,
+                                   BiConsumer<M, Supplier<NetworkEvent.Context>> messageConsumer) {
     INSTANCE.registerMessage(id++, messageType, encoder, decoder, messageConsumer);
   }
 }
