@@ -19,12 +19,12 @@
 
 package top.theillusivec4.culinaryconstruct.common.registry;
 
-import net.minecraft.block.Block;
-import net.minecraft.inventory.container.ContainerType;
-import net.minecraft.item.BlockItem;
-import net.minecraft.item.Item;
-import net.minecraft.item.ItemGroup;
-import net.minecraft.tileentity.TileEntityType;
+import net.minecraft.world.level.block.Block;
+import net.minecraft.world.inventory.MenuType;
+import net.minecraft.world.item.BlockItem;
+import net.minecraft.world.item.Item;
+import net.minecraft.world.item.CreativeModeTab;
+import net.minecraft.world.level.block.entity.BlockEntityType;
 import net.minecraftforge.common.extensions.IForgeContainerType;
 import net.minecraftforge.event.RegistryEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
@@ -33,7 +33,7 @@ import top.theillusivec4.culinaryconstruct.common.block.CulinaryStationBlock;
 import top.theillusivec4.culinaryconstruct.common.inventory.CulinaryStationContainer;
 import top.theillusivec4.culinaryconstruct.common.item.FoodBowlItem;
 import top.theillusivec4.culinaryconstruct.common.item.SandwichItem;
-import top.theillusivec4.culinaryconstruct.common.tileentity.CulinaryStationTileEntity;
+import top.theillusivec4.culinaryconstruct.common.blockentity.CulinaryStationBlockEntity;
 
 @Mod.EventBusSubscriber(bus = Mod.EventBusSubscriber.Bus.MOD)
 public class RegistryEventsHandler {
@@ -46,22 +46,22 @@ public class RegistryEventsHandler {
   @SubscribeEvent
   public static void onItemsRegistry(final RegistryEvent.Register<Item> evt) {
     BlockItem culinaryStation = new BlockItem(CulinaryConstructRegistry.CULINARY_STATION,
-        (new Item.Properties()).group(ItemGroup.DECORATIONS));
+        (new Item.Properties()).tab(CreativeModeTab.TAB_DECORATIONS));
     culinaryStation.setRegistryName(RegistryReference.CULINARY_STATION);
     evt.getRegistry().registerAll(culinaryStation, new SandwichItem(), new FoodBowlItem());
   }
 
   @SubscribeEvent
-  public static void onContainerRegistry(final RegistryEvent.Register<ContainerType<?>> evt) {
+  public static void onContainerRegistry(final RegistryEvent.Register<MenuType<?>> evt) {
     evt.getRegistry().register(IForgeContainerType.create(CulinaryStationContainer::new)
         .setRegistryName(RegistryReference.CULINARY_STATION));
   }
 
   @SuppressWarnings("ConstantConditions")
   @SubscribeEvent
-  public static void onTileEntityRegistry(final RegistryEvent.Register<TileEntityType<?>> evt) {
-    evt.getRegistry().register(TileEntityType.Builder
-        .create(CulinaryStationTileEntity::new, CulinaryConstructRegistry.CULINARY_STATION)
+  public static void onTileEntityRegistry(final RegistryEvent.Register<BlockEntityType<?>> evt) {
+    evt.getRegistry().register(BlockEntityType.Builder
+        .of(CulinaryStationBlockEntity::new, CulinaryConstructRegistry.CULINARY_STATION)
         .build(null).setRegistryName(RegistryReference.CULINARY_STATION));
   }
 }
